requirements.txt
httpx[http2]==0.27.0
aiosqlite==0.20.0
pyyaml==6.0.1
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.2.2
